+++ src/components/common/Layout/Layout.tsx
@@ -3,8 +3,8 @@ import {GlobalStyle} from 'styles/global-styles';
 import {lightTheme} from 'styles/theme';
 
 import type {ReactElement} from 'react';
-import {Header} from 'components/common/Header/Header';
-import {Footer} from 'components/common/Footer/Footer';
+import {Header} from 'components/Header/Nav';
+import {Footer} from 'components/Footer/Footer';
 
 interface LayoutProps {
   children?: ReactElement;
